% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_match.R
\name{direct_match}
\alias{direct_match}
\title{Direct Match Species & Genus Binomial}
\usage{
direct_match(df)
}
\arguments{
\item{df}{tibble containing the species binomial split into two columns: 'Genus' & 'Species'}
}
\value{
Returns a \code{tibble} with the same number of rows as the input \code{df} and with one additional Boolean column
\emph{New.Direct_Match} indicating whether the binomial was directly matched (TRUE) or not (FALSE)
}
\description{
Tries to directly match Genus + Species Binomial to Trees database. Does not perform any fuzzy matching nor is capturing any other spelling errors.
Expects a \code{tibble} input with species binomials split into the two columns 'Genus' and 'Species'.
}
\examples{
test1 \%>\% direct_match()
}
