% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_match_species_within_genus.R
\name{direct_match_species_within_genus}
\alias{direct_match_species_within_genus}
\title{Direct Match Species within Genus}
\usage{
direct_match_species_within_genus(df, backbone = NULL)
}
\arguments{
\item{df}{\code{tibble} containing the species binomial split into the columns \code{Orig.Genus} and \code{Orig.Species}.}

\item{backbone}{specifies which backbone is used: needs to be a subset of \code{c('BGCI', 'WCVP', 'WFO', 'GBIF', 'FIA', 'PM')} or \code{NULL} if the whole database should be used.}
}
\value{
Returns a \code{tibble} with the additional logical column \code{direct_match_species_within_genus}, indicating whether the specific epithet was successfully matched within the matched genus (TRUE) or not (FALSE).
}
\description{
Tries to directly match the specific epithet within an already matched genus in \code{Treemendous.Trees}
}
\examples{
iucn \%>\% dplyr::mutate(Matched.Genus = Orig.Genus) \%>\% direct_match_species_within_genus()
}
