% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fia}
\alias{fia}
\title{Cleaned Master Tree Species list from FIA}
\format{
A data frame with 2171 rows and 2 variables:
\describe{
\item{Genus}{Genus name of species binomial}
\item{Species}{Specific epithet of the species binomial}
}
}
\usage{
fia
}
\description{
A cleaned dataset containing trees the Forest Inventory and Analysis (FIA) program of the U.S. Forest Service. This dataset is used in the example usage section of the manuscript for the \emph{Treemendous} package.
The data was downloaded in November 2022 from the official webpage of the Forest Inventory and Analysis National Program and is available under the following \href{https://www.fia.fs.usda.gov/library/field-guides-methods-proc/index.php}{link}.
}
\keyword{datasets}
