% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match_species_within_genus.R
\name{fuzzy_match_species_within_genus}
\alias{fuzzy_match_species_within_genus}
\title{Fuzzy Match Specific Epithet within Genus}
\usage{
fuzzy_match_species_within_genus(df)
}
\arguments{
\item{df}{tibble containing the species binomial split into two columns: 'Genus' & 'Species'}
}
\value{
Returns a \code{tibble} with the same number of rows as the input \code{df} and with one additional Boolean column
\emph{New.fuzzy_match_species_within_genus} indicating whether the specific epithet was successfully fuzzy matched (TRUE) or not (FALSE)
}
\description{
Tries to fuzzy match the specific epithet within the same Genus to the Trees Database. Uses \code{fuzzyjoin::stringdist()} method to perform fuzzy matching.
}
\examples{
fuzzy_match_species_within_genus(test3)
}
