% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_flags.R
\name{highlight_flags}
\alias{highlight_flags}
\title{Highlight potential ambiguities introduced upon resolving synonyms to their accepted names}
\usage{
highlight_flags(df, backbone = NULL)
}
\arguments{
\item{df}{\code{tibble} which is the output of \code{resolve_synonyms()} and therefor contains the columns \code{Accepted.Genus}, \code{Accepted.Species}, \code{Matched.Genus} and \code{Matched.Species}. May contain additional columns, which will be discarded.}

\item{backbone}{specifies for which backbone(s) the flags should be appended. Needs to be one of (or a combination of) \code{c('WCVP', 'WFO', 'GBIF')} or \code{NULL} if all should be used.}
}
\value{
The function outputs a \code{tibble} that includes the original species names, the matched species names, the accepted species names, and the corresponding flags for the matched species.
It filters and returns only those entries where at least one flag was raised.
}
\description{
The user can call \code{highlight_flags()} from the output of \code{resolve_synonyms()} to investigate potential ambiguities when resolving synonyms to their accepted latin binomial names.
These ambiguities can be of varying importance for different use-cases and should be carefully assessed for each use-case.
}
\details{
The following flags have been introduced in order to clarify and alleviate potential ambiguities that were introduced upon compilation of \code{Treemendous.Trees}.
The flags always correspond to the latin binomial in the columns \code{c(Matched.Genus, Matched.Species)}.
The user is encouraged to check for their individual use-cases if she/he wants to exclude resolved names based on these flags. Below they are described in more detail.

The first column appended has the suffix \verb{_authorship_ambiguity} and can be either \code{TRUE} or \code{FALSE}.
\itemize{
\item \code{TRUE}: For a given latin binomial,
there are multiple entries at taxonomic rank "Species" in the underlying backbone,
which would be resolved to different latin binomials due to different authorships.
\itemize{
\item For instance, in the \code{WCVP} backbone, two entries are present for the latin binomial \emph{Acer flabellatum} at rank "Species" ( \emph{Acer flabellatum} Greene, \emph{Acer flabellatum} Rehder).
The first is considered a synonym of \emph{Acer macrophyllum} Pursh, while the latter is considered a synonym of \emph{Acer campbellii subsp. flabellatum} (Rehder) A.E.Murray.
}
\item \code{FALSE}: no ambiguity
}

The second column appended has the suffix \verb{_infraspecific_ambiguity} and can be either \code{TRUE} or \code{FALSE}.
\itemize{
\item \code{TRUE}: For a given latin binomial,
there are multiple entries at different taxonomic ranks (Species, Subspecies, Variety or Form) in the underlying backbone,
which would be resolved to different latin binomials.
\itemize{
\item For instance, in the \code{WCVP} backbone, four entries are present for the latin binomial \emph{Nothofagus obliqua}, one at rank "Species" ( \emph{Nothofagus obliqua}),
two at rank "Subspecies" ( \emph{Nothofagus obliqua subsp. andina} and \emph{Nothofagus obliqua subsp. valdiviana}) and one at rank "Variety" ( \emph{Nothofagus obliqua var. macrocarpa}).
While the former three would all be resolved to the latin binomial \emph{Nothofagus obliqua}, the latter is considered to be a synonym of the accepted name \emph{Nothofagus macrocarpa}.
\code{Resolve_synonyms()} will however always resolve it to \emph{Nothofagus obliqua}, because this was the only accepted name out of all the four entries.
This flag should not be a problem if you are working with latin binomials from the beginning.
However, if your initial species names consists of trinomials (infraspecific species names), then this flag can help you to identify ambiguous name resolving.
}
\item \code{FALSE}: no ambiguity
}

The third column appended has the suffix \verb{_infraspecific_link} and can be either \code{TRUE} or \code{FALSE}.
\itemize{
\item \code{TRUE}: indicates the the resolved species was obtained by linking a trinomial to a corresponding Latin binomial.
If an entry in a given backbone is pointing (meaning is a synonym of) to another entry which has been removed (in favor of another entry with the same latin binomial),
then the linkage of the synonym-accepted relation was updated. If you are only interested in resolving to the correct latin binomials, you can generally neglect this flag.
However, be aware that the associated information of the resolved species (e.g. Authorship information) might not be correct in your case.
\itemize{
\item For instance, \code{WCVP} considers the species \emph{Betula kwangsiensis} as a synonym of the accepted subspecies \emph{Betula kweichowensis subsp. kweichowensis}.
This subspecies however shares its latin binomial with the accepted species \emph{Betula kweichowensis}.
Therefore the species \emph{Betula kwangsiensis} had to be relinked to the species \emph{Betula kweichowensis} in order to resolve the latin binomials correctly.
\item Another example would be: According to \code{WFO}, the species \emph{Abies shastensis} is a synonym of the accepted variety \emph{Abies magnifica var. shastensis}.
However, because there is also an accepted entry at rank species \emph{Abies magnifica} (which is selected during the compilation of \code{Treemendous.Trees}),
the species \emph{Abies shastensis} has to be relinked to \emph{Abies magnifica}.
}
\item \code{FALSE}: no new linkaged was required
}
}
\examples{
iucn_resolved <- iucn \%>\% matching('WCVP') \%>\% resolve_synonyms('WCVP')
iucn_resolved \%>\% highlight_flags('WCVP')
}
