% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_synonym_resolving.R
\name{iterate_synonym_resolving}
\alias{iterate_synonym_resolving}
\title{Resolves Synonyms and returns Accepted Species}
\usage{
iterate_synonym_resolving(
  df,
  first_backbone = "WFO",
  second_backbone = "WCVP",
  third_backbone = "GBIF"
)
}
\arguments{
\item{df}{tibble with species names matched to Trees.Full.}

\item{first_backbone}{default backbone}

\item{second_backbone}{asdf}

\item{third_backbone}{asdf}
}
\value{

}
\description{
Resolves synonyms iteratively according to the backbones \code{WFO}, \code{WCVP} or \code{GBIF}
}
\examples{

}
