% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iucn}
\alias{iucn}
\title{Title: TODO.}
\format{
A data frame with 384 rows and 2 variables:
\describe{
\item{Genus}{Genus name of species binomial}
\item{Species}{Specific epithet of species binomial}
\item{IUCN_Family}{Taxonomical information on family-level}
}
\usage{
iucn
}
\description{
A dataset containing TODO.
}
\keyword{datasets}
