% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_name.R
\name{match_name}
\alias{match_name}
\title{Match User Species Binomials to database Trees.Full}
\usage{
match_name(df)
}
\arguments{
\item{df}{}
}
\value{
returns a data.frame with the original species binomials as wells as the matched species binomials from the Trees.Full database
}
\description{
\code{match_name()} tries to match every input species to a species in the Trees.Full database. First, we try to match the genus name followed by the species epithet.
}
\examples{
# match to a subset of Trees.Full
match_name(test1)
# match to a modified subset of Trees.Full
match_name(test2)
}
