% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_synonyms.R
\name{resolve_synonyms}
\alias{resolve_synonyms}
\title{Resolve Synonyms for Matched Species Names}
\usage{
resolve_synonyms(df)
}
\arguments{
\item{df}{: tibble containing the two columns \code{New.Genus} and \code{New.Species}, which need to be created by calling \code{matching()}.}
}
\value{
tibble with two new columns: \code{Accepted.Genus} and \code{Accepted.Species}
}
\description{
This function is called after \code{matching()} and resolve synonyms based on the database \code{Trees.Full}.
Information on synonyms comes from the databases \code{WCVP}, \code{WFO} and \code{GBIF}. \code{WFO} is considered to be the primary backbone, \code{WFO} the secondary, and \code{GBIF} the tertiary.
}
\examples{
test6 \%>\% matching() \%>\% resolve_synonyms()
}
