% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_matching.R
\name{sequential_matching}
\alias{sequential_matching}
\title{sequentially match all according to specific backbone ordering}
\usage{
sequential_matching(df, sequential_backbones = NULL)
}
\arguments{
\item{sequential_backbones}{}
}
\value{

}
\description{
sequentially match all according to specific backbone ordering
}
