% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_within_genus_match.R
\name{species_within_genus_match}
\alias{species_within_genus_match}
\title{Match Specific Epithet}
\usage{
species_within_genus_match(df, backbone = NULL)
}
\arguments{
\item{df}{tibble containing the species binomial split into two columns: 'Genus' & 'Species'}
}
\value{
Returns a \code{tibble} with the same number of rows as the input \code{df} and with one additional Boolean column
\emph{Matched.species_within_genus_match} indicating whether the genus was matched (TRUE) or not (FALSE)
}
\description{
#' Tries to match the specific epithet within a Genus to the Trees database. Does not perform any fuzzy matching nor is capturing any other spelling errors.
Expects a \code{tibble} input with species binomials split into the two columns 'Genus' and 'Species'.
}
\examples{
test2 \%>\% dplyr::mutate(Matched.Genus = Orig.Genus) \%>\% species_within_genus_match()
}
