% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_output.R
\name{summarize_output}
\alias{summarize_output}
\title{Summarizes the output of the treemendous pipeline}
\usage{
summarize_output(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Summarizes the output of the treemendous pipeline
}
