% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test1}
\alias{test1}
\title{Subset of Trees.Full, which is used as a test-set.}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{Species}{Name of specific epithet}
\item{Genus}{Name of Genus}
}
}
\source{
\url{https://github.com/speckerf/}
}
\usage{
test1
}
\description{
A test dataset containing 100 randomly selected Species from the database Trees.Full.
The following modifications were performed
There were no modifications on the names performed.
}
\keyword{datasets}
