% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test3}
\alias{test3}
\title{Modified subset of Trees.Full, which is used as a test-set.}
\format{
A data frame with 10 rows and 2 variables:
\describe{
\item{Orig.Species}{Name of specific epithet}
\item{Orig.Genus}{Name of Genus}
}
}
\source{
\url{https://github.com/speckerf/}
}
\usage{
test3
}
\description{
A test dataset containing 10 Species, the first 5 from test1 and the first 5 from test2.
}
\keyword{datasets}
