% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test4}
\alias{test4}
\title{Modified subset of Trees.Full, which is used as a test-set.}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{Orig.Species}{Name of specific epithet}
\item{Orig.Genus}{Name of Genus}
}
}
\source{
\url{https://github.com/speckerf/}
}
\usage{
test4
}
\description{
A test dataset containing 100 randomly selected Species from the database.
The following modifications were performed:
\itemize{
\item The last character of the Genus column was removed and the Genus names should therefore be fuzzy matched.
}
}
\keyword{datasets}
