% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test5}
\alias{test5}
\title{Modified subset of Trees.Full, which is used as a test-set.}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{Species}{Name of specific epithet}
\item{Genus}{Name of Genus}
}
}
\source{
\url{https://github.com/speckerf/}
}
\usage{
test5
}
\description{
A test dataset containing 100 randomly selected Species from the database.
The following modifications were performed:
\itemize{
\item The last character of the Genus column was removed and the Genus names should therefore be fuzzy matched.
\item The last character of the Species column was removed, and therefore should be fuzzy matched.
}
}
\keyword{datasets}
