% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test6}
\alias{test6}
\title{Modified subset of Trees.Full, which is used as a test-set.}
\format{
A data frame with 40 rows and 2 variables:
\describe{
\item{Orig.Species}{Name of specific epithet}
\item{Orig.Genus}{Name of Genus}
}
}
\source{
\url{https://github.com/speckerf/}
}
\usage{
test6
}
\description{
A test dataset the first 10 species from \code{test1}, \code{test2}, \code{test4} and \code{test5}.
The following modifications were performed:
\itemize{
\item 10: unchanged
\item 10: last character from species removed
\item 10: last character from genus removed
\item 10: last character from both genus and species removed
}
}
\keyword{datasets}
