% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Trees.Full}
\alias{Trees.Full}
\title{Title: TODO.}
\format{
A data frame with 331414 rows and 27 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
\url{https://github.com/speckerf/}
}
\usage{
Trees.Full
}
\description{
A dataset containing TODO.
}
\keyword{datasets}
