% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Trees.Reduced}
\alias{Trees.Reduced}
\title{Prices of 50,000 round cut diamonds TODO.}
\format{
A data frame with 331414 rows and 10 variables:
\describe{
\item{TODO}{price, in US dollars}
\item{TODO}{weight of the diamond, in carats}
...
}
}
\source{
\url{https://github.com/speckerf/}
}
\usage{
Trees.Reduced
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds TODO.
}
\keyword{datasets}
