% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_match.R
\name{direct_match}
\alias{direct_match}
\title{Direct Match Species & Genus Binomial}
\usage{
direct_match(df, backbone = NULL, target_df = NULL)
}
\arguments{
\item{df}{\code{tibble} containing the species binomial split into the columns \code{Orig.Genus} and \code{Orig.Species}.}

\item{backbone}{specifies which backbone is used: needs to be a subset of \code{c('BGCI', 'WCVP', 'WFO', 'GBIF')} or \code{NULL} if the whole database should be used.}

\item{target_df}{is used if the user wants to provide a custom target dataset. The parameter is intended only for compatibility with the function translate_trees and not for direct usage.}
}
\value{
Returns a \code{tibble} with the additional logical column \code{direct_match}, indicating whether the binomial was successfully matched (TRUE) or not (FALSE)
}
\description{
Tries to directly match Genus + Species Binomial to \code{Treemendous.Trees}.
}
\examples{
iucn \%>\% direct_match()
}
