% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_matching.R
\name{enforce_matching}
\alias{enforce_matching}
\title{Title}
\usage{
enforce_matching(
  df,
  backbone,
  extended_synonym_search = T,
  relaxed_fuzzy_matching = T
)
}
\arguments{
\item{df}{asdf}

\item{backbone}{asdf}

\item{extended_synonym_search}{asdf}

\item{relaxed_fuzzy_matching}{asdf}
}
\value{

}
\description{
Title
}
\examples{
iucn \%>\% matching('WFO') \%>\% enforce_matching('WFO')
}
