% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match_genus.R
\name{fuzzy_match_genus}
\alias{fuzzy_match_genus}
\title{Fuzzy Match Genus Name}
\usage{
fuzzy_match_genus(df, backbone = NULL)
}
\arguments{
\item{df}{tibble containing the species binomial split into two columns: 'Genus' & 'Species'}

\item{backbone}{specifies which backbone is used: needs to be a subset of c('BGCI', 'WCVP', 'WFO', 'GBIF', 'FIA', 'PM') or NULL if the whole database should be used#'}
}
\value{
Returns a \code{tibble} with the same number of rows as the input \code{df} and with one additional Boolean column
\emph{Matched.fuzzy_match_genus} indicating whether the genus name was successfully fuzzy matched (TRUE) or not (FALSE)
}
\description{
Tries to fuzzy match the specific epithet within the same Genus to the Trees Database. Uses \code{fuzzyjoin::stringdist()} method to perform fuzzy matching.
}
\examples{
test4 \%>\% fuzzy_match_genus()
}
