% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match_genus.R
\name{fuzzy_match_genus}
\alias{fuzzy_match_genus}
\title{Fuzzy Match Genus Name}
\usage{
fuzzy_match_genus(df, backbone = NULL)
}
\arguments{
\item{df}{\code{tibble} containing the species binomial split into the columns \code{Orig.Genus} and \code{Orig.Species}.}

\item{backbone}{specifies which backbone is used: needs to be a subset of \code{c('BGCI', 'WCVP', 'WFO', 'GBIF', 'FIA', 'PM')} or \code{NULL} if the whole database should be used.}
}
\value{
Returns a \code{tibble} with the additional logical column \code{fuzzy_match_genus}, indicating whether the genus was successfully matched (TRUE) or not (FALSE).
Further, the additional column \code{fuzzy_genus_dist} returns the distance for every match.
}
\description{
Tries to fuzzy match the genus name to \code{Treemendous.Trees}. Uses \code{fuzzyjoin::stringdist()} to perform fuzzy matching.
}
\examples{
iucn \%>\%
    dplyr::mutate(Orig.Genus = stringr::str_replace(Orig.Genus, '.{1}$', '')) \%>\%
    fuzzy_match_genus()
}
