% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match_species_within_genus.R
\name{fuzzy_match_species_within_genus}
\alias{fuzzy_match_species_within_genus}
\title{Fuzzy Match Species within Genus}
\usage{
fuzzy_match_species_within_genus(df, backbone = NULL, target_df = NULL)
}
\arguments{
\item{df}{\code{tibble} containing the species binomial split into the columns \code{Orig.Genus} and \code{Orig.Species}.}

\item{backbone}{specifies which backbone is used: needs to be a subset of \code{c('BGCI', 'WCVP', 'WFO', 'GBIF')} or \code{NULL} if the whole database should be used.}

\item{target_df}{is used if the user wants to provide a custom target dataset. The parameter is intended only for compatibility with the function translate_trees and should not be directly used.}
}
\value{
Returns a \code{tibble} with the additional logical column \code{fuzzy_match_species_within_genus}, indicating whether the specific epithet was successfully fuzzy matched within the matched genus (TRUE) or not (FALSE).
}
\description{
Tries to fuzzy match the species name to \code{Treemendous.Trees} within a genus. Uses \code{fuzzyjoin::stringdist()} to perform fuzzy matching.
}
\examples{
iucn \%>\%
    dplyr::mutate(Orig.Genus = stringr::str_replace(Orig.Genus, '.{1}$', '')) \%>\%
    dplyr::mutate(Matched.Genus = Orig.Genus) \%>\%
    fuzzy_match_species_within_genus()
}
