% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genus_match.R
\name{genus_match}
\alias{genus_match}
\title{Match Genus name}
\usage{
genus_match(df, backbone = NULL, target_df = NULL)
}
\arguments{
\item{df}{\code{tibble} containing the species binomial split into the columns \code{Orig.Genus} and \code{Orig.Species}.}

\item{backbone}{specifies which backbone is used: needs to be a subset of \code{c('BGCI', 'WCVP', 'WFO', 'GBIF')} or \code{NULL} if the whole database should be used.}

\item{target_df}{is used if the user wants to provide a custom target dataset. The parameter is intended only for compatibility with the function translate_trees and should not be directly used.}
}
\value{
Returns a \code{tibble} with the additional logical column \code{genus_match}, indicating whether the genus was successfully matched (TRUE) or not (FALSE)
}
\description{
#' Tries to match the genus name to \code{Treemendous.Trees}.
}
\examples{
iucn \%>\% genus_match()
}
