% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iucn}
\alias{iucn}
\title{Selected Trees from the IUCN red list of Threatened Species}
\format{
A data frame with 384 rows and 2 variables:
\describe{
\item{Orig.Genus}{Genus name of species binomial}
\item{Orig.Species}{Specific epithet of the species binomial}
}
}
\usage{
iucn
}
\description{
A dataset containing threatened tree species from the genus \code{Acer} and the families \code{Betulaceae}, \code{Nothofagaceae} and \code{Theaceae}.
The data was downloaded in June 2022 from the official webpage of the International Union for Conservation of Nature (IUCN) and is available under the following \href{https://www.iucnredlist.org/resources/spatial-data-download}{link}.
}
\keyword{datasets}
