% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{matching}
\alias{matching}
\title{Runs whole matching pipeline}
\usage{
matching(df)
}
\arguments{
\item{df}{\code{tibble} containing the species binomial split into two columns: 'Genus' & 'Species'. May contain additional columns, which will be ignored and returned by the function.}
}
\value{
Returns a \code{tibble}
}
\description{
This function is the core of the \code{treemendous} package and a wrapper around the following functions:
\itemize{
\item \code{direct_match()}
\item \code{genus_match()}
\item \code{fuzzy_match_genus()}
\item \code{species_within_genus_match()}
\item \code{suffix_match_species_within_genus()}
\item \code{fuzzy_match_species_within_genus()}
}
}
\examples{
matching(test1)
}
