% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_synonyms.R
\name{resolve_synonyms}
\alias{resolve_synonyms}
\title{Resolve Synonyms for Matched Species Names}
\usage{
resolve_synonyms(df, backbones = NULL)
}
\arguments{
\item{df}{: tibble containing the two columns \code{Matched.Genus} and \code{Matched.Species}, which need to be created by calling \code{matching()}.}

\item{backbones}{specifies the order in which synonyms are resolved: needs to be a subset of c('BGCI', 'WCVP', 'WFO', 'GBIF') or NULL if the default ordering c('BGCI', 'WFO', 'WCVP', 'FIA', 'PM', 'GBIF') should be used .}
}
\value{
tibble with two new columns: \code{Accepted.Genus} and \code{Accepted.Species}
}
\description{
This function is called after \code{matching()} and resolve synonyms based on the database \code{Trees.Full}.
Information on synonyms comes from the databases \code{WCVP}, \code{WFO} and \code{GBIF}. \code{WFO} is considered to be the primary backbone, \code{WFO} the secondary, and \code{GBIF} the tertiary.
}
\examples{
backbones = c('BGCI', 'WFO')
iucn \%>\% matching(backbones) \%>\% resolve_synonyms(backbones)
}
