% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix_match_species_within_genus.R
\name{suffix_match_species_within_genus}
\alias{suffix_match_species_within_genus}
\title{Suffix Match Specific Epithet within Genus}
\usage{
suffix_match_species_within_genus(df)
}
\arguments{
\item{df}{tibble containing the species binomial split into two columns: 'Genus' & 'Species'}
}
\value{
Returns a \code{tibble} with the same number of rows as the input \code{df} and with one additional Boolean column
\emph{New.suffix_match_species_within_genus} indicating whether the specific epithet was successfully suffix matched (TRUE) or not (FALSE)
}
\description{
Tries to match specific epithet within the same Genus to the Trees Database by exchanging common suffixes. The list of common suffixes is: \code{c("a", "i", "is", "um", "us", "ae", "oides", "escens")}
}
\examples{
test3 \%>\% dplyr::mutate(Matched.Genus = Orig.Genus) \%>\% suffix_match_species_within_genus()
}
