% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix_match_species_within_genus.R
\name{suffix_match_species_within_genus}
\alias{suffix_match_species_within_genus}
\title{Suffix Match Species within Genus}
\usage{
suffix_match_species_within_genus(df, backbone = NULL, target_df = NULL)
}
\arguments{
\item{df}{\code{tibble} containing the species binomial split into the columns \code{Orig.Genus} and \code{Orig.Species}.}

\item{backbone}{specifies which backbone is used: needs to be a subset of \code{c('BGCI', 'WCVP', 'WFO', 'GBIF')} or \code{NULL} if the whole database should be used.}

\item{target_df}{is used if the user wants to provide a custom target dataset. The parameter is intended only for compatibility with the function translate_trees and should not be directly used.}
}
\value{
Returns a \code{tibble} with the additional logical column \code{suffix_match_species_within_genus}, indicating whether the specific epithet was successfully matched within the matched genus (TRUE) or not (FALSE).
}
\description{
Tries to match the specific epithet by exchanging common suffixes within an already matched genus in \code{Treemendous.Trees}. The following suffixes are captured: \code{c("a", "i", "is", "um", "us", "ae")}
}
\examples{
# substitute endings c('um$|i$|is$|us$|ae$') with 'a' of specific epithet
iucn_modified<- iucn \%>\%
    dplyr::mutate(Orig.Species = stringr::str_replace(Orig.Species, 'um$|i$|is$|us$|ae$', 'a'))
iucn_modified \%>\%
    dplyr::mutate(Matched.Genus = Orig.Genus) \%>\%
    suffix_match_species_within_genus(backbone = c('BGCI', 'WFO'))
}
