% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_output.R
\name{summarize_output}
\alias{summarize_output}
\title{Summarizes the output of the treemendous pipeline}
\usage{
summarize_output(df)
}
\arguments{
\item{df}{: \code{tibble} being the output of matching()/sequential_matching()/enforce_matching() and optionally resolve_synonyms().}
}
\value{
Returns a \code{list} containing summary information about the matched species names and if provided also the resolved species names.
}
\description{
Summarizes the output of the treemendous pipeline
}
\examples{
iucn \%>\% matching() \%>\% resolve_synonyms() \%>\% summarize_output()
}
