% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test6}
\alias{test6}
\title{Modified subset of Trees.Full, which is used as a test-set.}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{Species}{Name of specific epithet}
\item{Genus}{Name of Genus}
}
}
\source{
\url{https://github.com/speckerf/}
}
\usage{
test6
}
\description{
A test dataset containing 10 randomly selected Species from the database.
The following modifications were performed:
\itemize{
\item 2: unchanged
\item 2: one character typo in species
\item 2: changed suffix
\item 2: Typo in Genus
\item 2: Typo in Genus and Species name
}
}
\keyword{datasets}
