% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_trees.R
\name{translate_trees}
\alias{translate_trees}
\title{Translate species names according to a custom target database.}
\usage{
translate_trees(df, target, max_iter = 3)
}
\arguments{
\item{df}{\code{tibble} with species that the user wants to translate into the species names of target. Species binomial split into the columns \code{Genus} and \code{Species}}

\item{target}{\code{tibble} with a new custom target database. Species binomial split into the columns \code{Genus} and \code{Species}}

\item{max_iter}{parameter which is passed to \code{enforce_matching()} and controls the maximum depth for matches.}
}
\value{
Returns a \code{tibble} with the species names of the input \code{df} in \code{Orig.Genus}, \code{Orig.Species}, and the translated names in \code{Matched.Genus} and \code{Matched.Species}.
Process information from calling matching() and enforce_matching() is added to the output.
}
\description{
The function is essentially a wrapper around the functions \code{\link[=matching]{matching()}} and \code{\link[=enforce_matching]{enforce_matching()}}.
Species names from \code{df} are first directly matched to \code{target} by calling \code{matching(df, backbone = 'CUSTOM', target_df = target)}.
Subsequently, the function calls \code{enforce_matching(df, backbone = 'CUSTOM', target_df = target)} to increase the number of translated species.
}
\examples{
translate_trees(df = iucn, target = fia)
}
